requirements.txt
discord.py>=2.3.2,<3.0.0
Flask>=2.2.5,<3.0.0
openai>=1.0.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0